+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/autocaptions/AutoCaptionsPatch.kt
@@ -29,6 +29,7 @@ val autoCaptionsPatch = bytecodePatch(
             "19.34.42",
             "19.43.41",
             "19.45.38",
+            "19.46.42",
         ),
     )