+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/hide/player/flyoutmenupanel/HidePlayerFlyoutMenuPatch.kt
@@ -31,6 +31,7 @@ val hidePlayerFlyoutMenuPatch = bytecodePatch(
             "19.34.42",
             "19.43.41",
             "19.45.38",
+            "19.46.42",
         ),
     )